Issue #4142: Split and Organize Checkstyle inputs by Test for InterfaceTypeParameterName
